Function zng_wgt_scroll::cmd::scroll_to

pub fn scroll_to(
    target: impl ScrollToTargetProvider,
    mode: impl Into<ScrollToMode>,
)

Scroll all parent is_scroll widgets of target so that it becomes visible.

This function is a helper for searching for the target in all windows and sending SCROLL_TO_CMD for all required scroll widgets. Does nothing if the target is not found.