Function zng_wgt_scroll::auto_hide_extra

pub fn auto_hide_extra(
    child: impl UiNode,
    extra: impl IntoVar<SideOffsets>,
) -> impl UiNode

P Extra space added to the viewport auto-hide rectangle.

The scroll sets the viewport plus these offsets as the FrameBuilder::auto_hide_rect, this value is used for optimizations from the render culling to lazy widgets.

Scrolling can use only lightweight render updates to scroll within the extra margin, so there is an exchange of performance, a larger extra space means that more widgets are rendering, but also can mean less full frame requests, if there is no other widget requesting render.

By default is 500.dip().min(100.pct()), one full viewport extra capped at 500.

This property sets the AUTO_HIDE_EXTRA_VAR.