zng_wgt_scroll
Function auto_scroll_indicator

pub fn auto_scroll_indicator(
    child: impl UiNode,
    indicator: impl IntoVar<WidgetFn<AutoScrollArgs>>,
) -> impl UiNode

P Auto scroll icon/indicator node.

The indicator is instantiated on middle click if auto_scroll is enabled, the node is layered as an adorner of the scroll. All context vars and the full SCROLL context are captured and can be used in the indicator.

Is node::default_auto_scroll_indicator by default.