zng_wgt_scroll

Function zoom_scale

pub fn zoom_scale(child: impl IntoUiNode, scale: impl IntoVar<Factor>) -> UiNode

P Binds the zoom_scale scroll var to the property value.

The binding is bidirectional and the scroll variable is assigned on init.

Note that settings the offset directly overrides effects like smooth scrolling, prefer using the scroll commands to scroll over this property.