zng_wgt_scroll

Function zoom_touch_origin

pub fn zoom_touch_origin(
    child: impl UiNode,
    origin: impl IntoVar<Point>,
) -> impl UiNode

P Center point of zoom scaling done using the touch pinch gesture.

Relative values are resolved in the viewport space. The scroll offsets so that the point in the viewport and content stays as close as possible after the scale change.

The default value (Length::Default) is center point between the two touch contact points.

This property sets the ZOOM_TOUCH_ORIGIN_VAR.