zng_wgt_scroll::scrollbar

Function thumb

pub fn thumb(wgt: &mut WidgetBuilding, node: impl IntoUiNode)

B Thumb widget.

Recommended widget is Thumb!, but can be any widget that implements thumb behavior and tags itself in the frame.

Mixin Property

This property is a mixin, it modifies the widget builder during build, it does not produce a standalone UI node.