zng_wgt_scroll

Struct SCROLL

pub struct SCROLL;

Controls the parent scroll.

Implementations

impl SCROLL

pub fn try_id(&self) -> Option<WidgetId>

Gets the ID of the scroll ancestor represented by the SCROLL.

pub fn id(&self) -> WidgetId

Gets the ID of the scroll ancestor represented by the SCROLL.

Panics

Panics if not inside a scroll.

pub fn config_node(&self, child: impl IntoUiNode) -> UiNode

New node that holds data for the SCROLL context.

Scroll implementers must add this node to their context.

pub fn mode(&self) -> Var<ScrollMode>

Scroll mode of the parent scroll.

pub fn vertical_offset(&self) -> ContextVar<Factor>

Vertical offset of the parent scroll.

The value is a percentage of content.height - viewport.height.

This variable is usually read-write, but you should avoid modifying it directly as direct assign as the value is not validated and does not participate in smooths scrolling. Prefer the scroll methods of this service to scroll.

pub fn horizontal_offset(&self) -> ContextVar<Factor>

Horizontal offset of the parent scroll.

The value is a percentage of content.width - viewport.width.

This variable is usually read-write, but you should avoid modifying it directly as direct assign as the value is not validated and does not participate in smooths scrolling. Prefer the scroll methods of this service to scroll.

pub fn zoom_scale(&self) -> ContextVar<Factor>

Zoom scale factor of the parent scroll.

This variable is usually read-write, but you should avoid modifying it directly as direct assign as the value is not validated and does not participate in smooths scrolling. Prefer the zoom methods of this service to change scale.

pub fn rendered_offset(&self) -> Factor2d

Latest rendered offset.

pub fn rendered_zoom_scale(&self) -> Factor

Latest rendered zoom scale factor.

pub fn vertical_overscroll(&self) -> Var<Factor>

Extra vertical offset, requested by touch gesture, that could not be fulfilled because vertical_offset is already at 0.fct() or 1.fct().

The factor is between in the -1.0..=1.0 range and represents the overscroll offset in pixels divided by the viewport width.

pub fn horizontal_overscroll(&self) -> Var<Factor>

Extra horizontal offset requested that could not be fulfilled because horizontal_offset is already at 0.fct() or 1.fct().

The factor is between in the -1.0..=1.0 range and represents the overscroll offset in pixels divided by the viewport width.

pub fn vertical_ratio(&self) -> Var<Factor>

Ratio of the scroll parent viewport height to its content.

The value is viewport.height / content.height.

pub fn horizontal_ratio(&self) -> Var<Factor>

Ratio of the scroll parent viewport width to its content.

The value is viewport.width / content.width.

pub fn vertical_content_overflows(&self) -> Var<bool>

If the vertical scrollbar should be visible.

pub fn horizontal_content_overflows(&self) -> Var<bool>

If the horizontal scrollbar should be visible.

pub fn viewport_size(&self) -> Var<PxSize>

Latest computed viewport size of the parent scroll.

pub fn content_size(&self) -> Var<PxSize>

Latest computed content size of the parent scroll.

The size is scaled by the zoom scale.

pub fn content_original_size(&self) -> Var<PxSize>

Latest layout content size of the parent scroll. Not scaled by zoom.

pub fn scroll_vertical(&self, delta: ScrollFrom)

Applies the delta to the vertical offset.

If smooth scrolling is enabled it is used to update the offset.

pub fn scroll_horizontal(&self, delta: ScrollFrom)

Applies the delta to the horizontal offset.

If smooth scrolling is enabled the chase animation is created or updated by this call.

pub fn scroll_vertical_clamp(&self, delta: ScrollFrom, min: f32, max: f32)

Applies the delta to the vertical offset, but clamps the final offset by the inclusive min and max.

If smooth scrolling is enabled it is used to update the offset.

pub fn scroll_horizontal_clamp(&self, delta: ScrollFrom, min: f32, max: f32)

Applies the delta to the horizontal offset, but clamps the final offset by the inclusive min and max.

If smooth scrolling is enabled it is used to update the offset.

pub fn auto_scroll(&self, velocity: DipVector)

Animate scroll at the direction and velocity (in DIPs per second).

pub fn scroll_vertical_touch(&self, delta: Px)

Applies the delta to the vertical offset without smooth scrolling and updates the vertical overscroll if it changes.

This method is used to implement touch gesture scrolling, the delta is always ScrollFrom::Var.

pub fn scroll_horizontal_touch(&self, delta: Px)

Applies the delta to the horizontal offset without smooth scrolling and updates the horizontal overscroll if it changes.

This method is used to implement touch gesture scrolling, the delta is always ScrollFrom::Var.

pub fn clear_vertical_overscroll(&self)

Quick ease vertical overscroll to zero.

pub fn clear_horizontal_overscroll(&self)

Quick ease horizontal overscroll to zero.

pub fn scroll_vertical_touch_inertia(&self, delta: Px, duration: Duration)

Animates to delta over duration.

pub fn scroll_horizontal_touch_inertia(&self, delta: Px, duration: Duration)

Animates to delta over duration.

pub fn chase_vertical(&self, modify_offset: impl FnOnce(Factor) -> Factor)

Set the vertical offset to a new offset derived from the last, blending into the active smooth scrolling chase animation, or starting a new one, or just setting the var if smooth scrolling is disabled.

pub fn chase_horizontal(&self, modify_offset: impl FnOnce(Factor) -> Factor)

Set the horizontal offset to a new offset derived from the last set offset, blending into the active smooth scrolling chase animation, or starting a new one, or just setting the var if smooth scrolling is disabled.

pub fn chase_zoom(&self, modify_scale: impl FnOnce(Factor) -> Factor)

Set the zoom scale to a new scale derived from the last set scale, blending into the active smooth scaling chase animation, or starting a new or, or just setting the var if smooth scrolling is disabled.

pub fn zoom(&self, modify_scale: impl FnOnce(Factor) -> Factor, origin: PxPoint)

Zoom in or out keeping the origin point in the viewport aligned with the same point in the content.

pub fn zoom_touch( &self, phase: TouchPhase, scale: Factor, center_in_viewport: Point2D<f32, Px>, )

Applies the scale to the current zoom scale without smooth scrolling and centered on the touch point.

pub fn can_scroll_vertical(&self) -> Var<bool>

Gets a var that is true when the content height is greater then the viewport height.

pub fn can_scroll_horizontal(&self) -> Var<bool>

Gets a var that is true when the content width is greater then the viewport with.

pub fn can_scroll_down(&self) -> Var<bool>

Gets a var that is true when the content height is greater then the viewport height and the vertical offset is not at the maximum.

pub fn can_scroll_up(&self) -> Var<bool>

Gets a var that is true when the content height is greater then the viewport height and the vertical offset is not at the minimum.

pub fn can_scroll_left(&self) -> Var<bool>

Gets a var that is true when the content width is greater then the viewport width and the horizontal offset is not at the minimum.

pub fn can_scroll_right(&self) -> Var<bool>

Gets a var that is true when the content width is greater then the viewport width and the horizontal offset is not at the maximum.

pub fn scroll_to(&self, mode: impl Into<ScrollToMode>)

Scroll the WIDGET into view.

pub fn scroll_to_zoom( &self, mode: impl Into<ScrollToMode>, zoom: impl Into<Factor>, )

Scroll the WIDGET into view and adjusts the zoom scale.

pub fn can_zoom_in(&self) -> bool

Returns true if the content can be scaled and the current scale is less than the max.

pub fn can_zoom_out(&self) -> bool

Returns true if the content can be scaled and the current scale is more than the min.

pub fn actual_min_zoom(&self) -> Factor

Compute the minimum scale allowed.

This is the MIN_ZOOM_VAR unless it is set to < 0.01%, in that case it is the fit scale or 100% if the content is smaller than the viewport.

impl SCROLL

pub fn context_values_set(&self, set: &mut ContextValueSet)

Insert the context values used by SCROLL in the set.

Capturing this set plus all context vars enables using all SCROLL methods outside the scroll.