zng_wgt_scroll::thumb

Function viewport_ratio

pub fn viewport_ratio(ratio: impl IntoVar<Factor>)

c Viewport/content ratio.

This becomes the height for vertical and width for horizontal.

Capture-Only

This property is capture-only, it only defines a property signature, it does not implement any behavior by itself. Widgets can capture and implement this property as part of their intrinsics, otherwise it will have no effect if set on a widget that does not implement it.