zng_wgt_scroll

Module node

UI nodes used for building the scroll widget.

Functions

access_scroll_node

Create a node that converts ACCESS_SCROLL_EVENT to command requests.

auto_scroll_node

Create a note that spawns the auto scroller on middle click and fulfill AUTO_SCROLL_CMD requests.

default_auto_scroll_indicator

Renders a white circle with arrows that indicate what directions can be scrolled.

h_scrollbar_presenter

Create a node that generates and presents the horizontal scrollbar.

overscroll_node

Overscroll visual indicator.

page_commands_node

Create a node that implements PAGE_UP_CMD, PAGE_DOWN_CMD, PAGE_LEFT_CMD and PAGE_RIGHT_CMD scoped on the widget.

scroll_commands_node

Create a node that implements SCROLL_UP_CMD, SCROLL_DOWN_CMD, SCROLL_LEFT_CMD and SCROLL_RIGHT_CMD scoped on the widget.

scroll_to_edge_commands_node

Create a node that implements SCROLL_TO_TOP_CMD, SCROLL_TO_BOTTOM_CMD, SCROLL_TO_LEFTMOST_CMD and SCROLL_TO_RIGHTMOST_CMD scoped on the widget.

scroll_to_node

Create a node that implements SCROLL_TO_CMD scoped on the widget and scroll to focused.

scroll_touch_node

Create a node that implements scroll by touch gestures for the widget.

scroll_wheel_node

Create a node that implements scroll-wheel handling for the widget.

scrollbar_joiner_presenter

Create a node that generates and presents the scrollbar joiner.

v_scrollbar_presenter

Create a node that generates and presents the vertical scrollbar.

viewport

The actual content presenter.

zoom_commands_node

Create a node that implements ZOOM_IN_CMD, ZOOM_OUT_CMD, ZOOM_TO_FIT_CMD, and ZOOM_RESET_CMD scoped on the widget.